[qca-nss-ecm] Bridge vlan passthrough support from ecm
(cherry picked from commit 45507ddaf3875324eabf96f374d80c2038483474)

+++ frontends/sfe/ecm_sfe_common.c
@@ -229,7 +229,7 @@ fail:
  * Step 2.) Fast transmit setting is enabled on the destination interface only when no qdisc
  * 	is found in the hierarchy.
  */
-void ecm_sfe_common_fast_xmit_set(uint16_t *rule_flags, uint16_t *valid_flags, struct sfe_qdisc_rule *qdisc_rule, struct ecm_db_iface_instance *from_ifaces[ECM_DB_IFACE_HEIRARCHY_MAX], struct ecm_db_iface_instance *to_ifaces[ECM_DB_IFACE_HEIRARCHY_MAX], int32_t from_interfaces_first, int32_t to_interfaces_first)
+void ecm_sfe_common_fast_xmit_set(uint32_t *rule_flags, uint32_t *valid_flags, struct sfe_qdisc_rule *qdisc_rule, struct ecm_db_iface_instance *from_ifaces[ECM_DB_IFACE_HEIRARCHY_MAX], struct ecm_db_iface_instance *to_ifaces[ECM_DB_IFACE_HEIRARCHY_MAX], int32_t from_interfaces_first, int32_t to_interfaces_first)
 {
 	s32 interface_num;
 	bool qdisc_found = false;
